+++ example/example.dart
@@ -2,101 +2,162 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_reorderable_grid_view/reorderable_grid_view.dart';
 
-class ExamplePage extends StatelessWidget {
-  const ExamplePage({Key? key}) : super(key: key);
+/// Example how you should act when updating the children.
+///
+/// In this example, you can add, remove or clear all children.
+///
+/// Also there is a button to update the first child text to 999 to see that
+/// you still get the updated children that are added to the GridView.
+///
+/// If you reorder the children, you have an example how you should make use
+/// of the onUpdate function.
+///
+/// The black square is locked and can't be moved because it was added to
+/// lockedChildren.
+class HomePage extends StatefulWidget {
+  const HomePage({Key? key}) : super(key: key);
 
   @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Colors.grey,
-      body: Padding(
-        padding: const EdgeInsets.fromLTRB(20, 100, 20, 0),
-        child: Column(
-          children: [
-            Expanded(
-              child: ReorderableGridView(
-                spacing: 12,
-                children: List.generate(
-                  20,
-                  (index) => Container(
-                    color: Colors.blue,
-                    height: 100,
-                    width: 100,
-                    child: Text(
-                      'test $index',
-                      style: const TextStyle(
-                        fontSize: 20,
-                        color: Colors.white,
-                      ),
-                    ),
-                  ),
-                ),
-              ),
-            ),
-            Container(
-              color: Colors.red,
-              child: const SafeArea(
-                child: Text(
-                  'i bims 1 text',
-                  style: TextStyle(
-                    fontSize: 24,
-                    fontWeight: FontWeight.bold,
-                  ),
-                ),
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
+  State<HomePage> createState() => _HomePageState();
 }
 
-class ExampleWithLockedItemsPage extends StatelessWidget {
-  const ExampleWithLockedItemsPage({Key? key}) : super(key: key);
+class _HomePageState extends State<HomePage> {
+  List<int> children = <int>[];
 
   @override
   Widget build(BuildContext context) {
-    final lockedChildren = [0, 4];
+    final lockedChildren = <int>[1];
+
     return Scaffold(
       backgroundColor: Colors.grey,
       body: Padding(
-        padding: const EdgeInsets.fromLTRB(20, 100, 20, 0),
+        padding: const EdgeInsets.fromLTRB(20, 20, 20, 0),
         child: Column(
           children: [
-            Expanded(
-              child: ReorderableGridView(
-                lockedChildren: lockedChildren,
-                spacing: 12,
-                children: List.generate(
-                  20,
-                  (index) => Container(
-                    color: lockedChildren.contains(index)
-                        ? Colors.black
-                        : Colors.blue,
-                    height: 100,
-                    width: 100,
-                    child: Text(
-                      'test $index',
-                      style: const TextStyle(
-                        fontSize: 20,
-                        color: Colors.white,
+            Padding(
+              padding: const EdgeInsets.only(bottom: 20),
+              child: Row(
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: [
+                  GestureDetector(
+                    onTap: () {
+                      if (children.isNotEmpty) {
+                        children[0] = 999;
+                        setState(() {
+                          children = children;
+                        });
+                      }
+                    },
+                    child: Container(
+                      color: Colors.deepOrangeAccent,
+                      height: 100,
+                      width: 100,
+                      child: const Center(
+                        child: Icon(
+                          Icons.find_replace,
+                          size: 36,
+                        ),
                       ),
                     ),
                   ),
-                ),
+                  const SizedBox(width: 20),
+                  GestureDetector(
+                    onTap: () {
+                      setState(() {
+                        children = children..add(children.length);
+                      });
+                    },
+                    child: Container(
+                      color: Colors.green,
+                      height: 100,
+                      width: 100,
+                      child: const Center(
+                        child: Icon(
+                          Icons.add,
+                          size: 36,
+                        ),
+                      ),
+                    ),
+                  ),
+                  const SizedBox(width: 20),
+                  GestureDetector(
+                    onTap: () {
+                      if (children.isNotEmpty) {
+                        setState(() {
+                          children = children..removeLast();
+                        });
+                      }
+                    },
+                    child: Container(
+                      color: Colors.red,
+                      height: 100,
+                      width: 100,
+                      child: const Center(
+                        child: Icon(
+                          Icons.remove,
+                          size: 36,
+                        ),
+                      ),
+                    ),
+                  ),
+                  const SizedBox(width: 20),
+                  GestureDetector(
+                    onTap: () {
+                      if (children.isNotEmpty) {
+                        setState(() {
+                          children = <int>[];
+                        });
+                      }
+                    },
+                    child: Container(
+                      color: Colors.yellowAccent,
+                      height: 100,
+                      width: 100,
+                      child: const Center(
+                        child: Icon(
+                          Icons.delete,
+                          size: 36,
+                        ),
+                      ),
+                    ),
+                  ),
+                ],
               ),
             ),
-            Container(
-              color: Colors.red,
-              child: const SafeArea(
-                child: Text(
-                  'i bims 1 text',
-                  style: TextStyle(
-                    fontSize: 24,
-                    fontWeight: FontWeight.bold,
+            Expanded(
+              child: ReorderableGridView(
+                enableLongPress: false,
+                lockedChildren: lockedChildren,
+                spacing: 12,
+                onUpdate: (int oldIndex, int newIndex) {
+                  setState(() {
+                    final draggedItem = children[oldIndex];
+                    final collisionItem = children[newIndex];
+                    children[newIndex] = draggedItem;
+                    children[oldIndex] = collisionItem;
+                  });
+                },
+                children: [
+                  ...List.generate(
+                    children.length,
+                    (index) => Container(
+                      color: lockedChildren.contains(index)
+                          ? Colors.black
+                          : Colors.blue,
+                      height: 100,
+                      width: 100,
+                      child: Center(
+                        child: Text(
+                          'test ${children[index]}',
+                          style: const TextStyle(
+                            fontSize: 20,
+                            color: Colors.white,
+                          ),
+                        ),
+                      ),
+                    ),
                   ),
-                ),
+                ],
               ),
             ),
           ],